pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "send-to-kindle"
version = "0.1.0"
description = "Private service that converts article URLs into EPUB files and emails them to Kindle devices."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn[standard]>=0.34,<1.0",
  "httpx>=0.28,<1.0",
  "trafilatura>=2.0,<3.0",
  "ebooklib>=0.18",
  "aiosmtplib>=3.0,<4.0",
  "PyYAML>=6.0,<7.0",
  "beautifulsoup4>=4.12,<5.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
]

[project.scripts]
send-to-kindle = "send_to_kindle.main:main"

[tool.setuptools.packages.find]
include = ["send_to_kindle*"]
